✨ Player と Boss の共通処理を統合した Actor 基底クラスを作成
- Actor 基底クラスを新規作成し、共通のプロパティとメソッドを統合
- Player クラスを Actor 継承に変更、重複コードを削除
- Boss クラスを Actor 継承に変更、無限MP システムを実装
- 約200行の重複コードを削除し、保守性を向上
